[cargo-nextest] reorganize dispatch module
We did a reorganization a while ago but I wasn't satisfied by that. Let's try
again: split dispatch into core and utility modules.

This is a refactor with zero functional changes.
