+++ cargo-nextest/src/dispatch/core/mod.rs
@@ -1,0 +1,24 @@
+// Copyright (c) The nextest Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Core commands: list, run, bench, archive, replay.
+//!
+//! These commands share infrastructure (BaseApp, App, test filtering, cargo
+//! integration) and are the primary purpose of nextest.
+
+mod archive;
+mod base;
+mod filter;
+mod list;
+mod replay;
+mod run;
+#[cfg(test)]
+mod tests;
+mod value_enums;
+
+pub(crate) use archive::{ArchiveApp, ArchiveOpts};
+pub(crate) use base::{BaseApp, current_version};
+pub(crate) use filter::TestBuildFilter;
+pub(crate) use list::ListOpts;
+pub(crate) use replay::{ReplayOpts, exec_replay};
+pub(crate) use run::{App, BenchOpts, RunOpts};
