+++ cargo-nextest/src/dispatch/utility/mod.rs
@@ -1,0 +1,17 @@
+// Copyright (c) The nextest Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Utility commands: show-config, self, debug, store.
+//!
+//! These commands are management/debugging tools that are standalone and don't
+//! share core infrastructure.
+
+mod debug;
+mod self_cmd;
+mod show_config;
+mod store;
+
+pub(crate) use debug::{DebugCommand, ExtractOutputFormat};
+pub(crate) use self_cmd::SelfCommand;
+pub(crate) use show_config::ShowConfigCommand;
+pub(crate) use store::StoreCommand;
